registration form updated
